moved data gen to contactnet folder
